+++ server/page/BasePage.php
@@ -130,7 +130,7 @@ abstract class BasePage
             "/css/ext/query-builder.default.min.css",
             "/css/ext/query-builder.bootstrap.css",            
             "/css/ext/bootstrap4-toggle.min.css",      
-            "/css/ext/simplemde.min.css",   
+            "/css/ext/easymde.min.css",   
         );
         $this->js_includes = array(
             "/js/ext/jquery.min.js",
@@ -151,7 +151,7 @@ abstract class BasePage
             "/js/ext/moment.min.js",
             "/js/ext/query-builder.standalone.min.js",
             "/js/ext/bootstrap4-toggle.min.js",
-            "/js/ext/simplemde.min.js",
+            "/js/ext/easymde.min.js",
         );
         $this->js_includes_after = array(
             "/js/ext/loader.js",