+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - add option for `keyword` change
  - add option for `page_access_type` change
  - add option for page `is_headless` change
+ - add markdown editor for fields from type `markdown`
 
 ### Bugfix 
  - fix json condition builder for `in_one_of` type used with other condition